fixes merge errors
