+++ plugin/exo/Resources/modules/quiz/player/components/item-player.jsx
@@ -1,22 +1,62 @@
 import React, {PropTypes as T} from 'react'
 
-const Hints = () =>
-  <div>
-    This is the hint section
+import {tex, transChoice} from './../../../utils/translate'
+
+const UsedHint = props =>
+  <div className="well well-sm">
+    <span className="fa fa-fw fa-eye fa-lightbulb-o"></span>
+    {props.value}
+  </div>
+
+UsedHint.propType = {
+  value: T.string.isRequired
+}
+
+const Hint = props =>
+  <button type="button" className="btn btn-default btn-block">
+    <span className="fa fa-fw fa-eye"></span>
+    {tex('hint_show')}
+
+    {props.penalty &&
+    <small className="text-danger">
+      {transChoice('hint_penalty', props.penalty, {count: props.penalty}, 'ujm_exo')}
+    </small>
+    }
+  </button>
+
+Hint.propTypes = {
+  penalty: T.number,
+  showHint: T.func.isRequired
+}
+
+const Hints = props =>
+  <div className="item-hints">
+    {props.hints.map((hint, index) =>
+      <Hint
+        key={index}
+        penalty={hint.penalty}
+        showHint={() => props.showHint(hint)}
+      />
+    )}
   </div>
 
 Hints.propTypes = {
-  hints: T.array.isRequired
+  hints: T.arrayOf(T.shape({
+    id: T.string.isRequired,
+    penalty: T.number
+  })).isRequired,
+  showHint: T.func.isRequired
 }
 
-export const ItemPlayer = props =>
+const ItemPlayer = props =>
   <div className="item-player">
-    <header className="item-content">
-      <strong dangerouslySetInnerHTML={{__html: props.item.content}}/>
-    </header>
+    <p>{props.item.content}</p>
+
+    <hr/>
     {props.children}
+    <hr/>
     {props.item.hints && 0 !== props.item.hints.length &&
-      <Hints hints={props.item.hints}/>
+    <Hints hints={props.item.hints} showHint={props.showHint} />
     }
   </div>
 
@@ -25,5 +65,8 @@ ItemPlayer.propTypes = {
     content: T.string.isRequired,
     hints: T.array
   }).isRequired,
+  showHint: T.func.isRequired,
   children: T.node.isRequired
 }
+
+export {ItemPlayer}
